refactored test structure
