--- test/main_tester.cpp
@@ -1,8 +1,0 @@
-#include "fib.cpp"
-#include "fib_nobase.cpp"
-int main()
-{
-	fib_test();
-//	fibnobase_test();
-
-}
